grafana_upload_middleware/requirements.txt
fastapi
pydantic
uvicorn
pyserial
influxdb-client
